+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.4] - 2021-06-17
+
+### Changed
+
+- `AccessBanner.tsx` - update old hook references
+- `ArticleMeta.tsx` - update old hook references
+- `CategoryCard.tsx` - update old hook references
+- `CategoryList.tsx` - update old hook references
+
 ## [0.2.3] - 2021-04-28
 
 ### Added